+++ src/app/kid/register/chapter/Chapter5.tsx
@@ -1,70 +1,168 @@
 "use client";
 
-import React, { useEffect, useState } from "react";
+import React, { useState } from "react";
 import { Label } from "@/elements/label/Label";
 import Button from "@/elements/button/Button";
 import { ChapterProps } from "@/hook/useChapter";
-import styles from "../profile.module.css";
+import ListInput from "@/elements/input/ListInput";
+import Circleline from "@/elements/svg/Checkbox";
+import styles from "../kid.module.css";
+import XIcon from "@/elements/svg/XIcon";
 import { useProfileStore } from "@/store/useProfileStore";
-import { useModalStore } from "@/store/useModalStore";
 
 const Chapter5: React.FC<ChapterProps> = ({ onNext }) => {
-    const [etcInput, setEtcInput] = useState("");
-    const setEtc = useProfileStore((state) => state.setEtc);
-    const { openModal, setComp, closeModal } = useModalStore();
+    const [symptomInputValue, setSymptomInputValue] = useState("");
+    const [allergicInputValue, setAllergicInputValue] = useState("");
+    const [symptomList, setSymptomList] = useState<string[]>([]);
+    const [allergicList, setAllergicList] = useState<string[]>([]);
 
-    useEffect(() => {
-        setComp(
-            <>
-                <div className={styles.modalHeader}>
-                    <p>
-                        김아기 프로필
-                        <br />
-                        등록완료!
-                    </p>
-                </div>
-                <div className={styles.modalSubDesc}>
-                    <p>
-                        아이가 더 효과적인 치료를
-                        <br />
-                        받을 수 있게 도와드릴게요!
-                    </p>
-                </div>
-                <Button
-                    css="profileBtn"
-                    label="작성 완료"
-                    onClick={closeModal}
-                ></Button>
-            </>,
-        );
-    }, []);
+    const setSymptom = useProfileStore((state) => state.setSymptom);
+    const setAllergic = useProfileStore((state) => state.setAllergic);
+
+    // 증상 추가
+    const addSymptom = () => {
+        const trimmedValue = symptomInputValue.trim();
+        if (trimmedValue && !symptomList.includes(trimmedValue)) {
+            setSymptomList((prev) => [...prev, trimmedValue]);
+            setSymptomInputValue("");
+        }
+    };
+
+    // 증상 입력
+    const handleSymptomKeyDown = (e: React.KeyboardEvent<HTMLInputElement>) => {
+        // 한글 입력 중이면 무시
+        if (e.nativeEvent.isComposing) {
+            return;
+        }
+
+        if (e.key === "Enter") {
+            e.preventDefault();
+            addSymptom();
+        }
+    };
+
+    // 증상 삭제
+    const removeSymptom = (index: number) => {
+        setSymptomList((prev) => prev.filter((_, i) => i !== index));
+    };
+
+    // 알레르기 추가 함수
+    const addAllergic = () => {
+        const trimmedValue = allergicInputValue.trim();
+        console.log("입력된 값:", trimmedValue); // 입력값 확인
+
+        if (trimmedValue && !allergicList.includes(trimmedValue)) {
+            console.log("추가되기 전 리스트:", allergicList);
+            setAllergicList((prev) => {
+                console.log("업데이트될 리스트:", [...prev, trimmedValue]);
+                return [...prev, trimmedValue];
+            });
+            setAllergicInputValue("");
+        }
+    };
+
+    // 알레르기 입력
+    const handleAllergicKeyDown = (
+        e: React.KeyboardEvent<HTMLInputElement>,
+    ) => {
+        // 한글 입력 중이면 무시
+        if (e.nativeEvent.isComposing) {
+            return;
+        }
+
+        if (e.key === "Enter") {
+            e.preventDefault();
+            addAllergic();
+        }
+    };
+    // 알레르기 삭제
+    const removeAllergic = (index: number) => {
+        setAllergicList((prev) => prev.filter((_, i) => i !== index));
+    };
 
     const handleNext = () => {
-        setEtc(etcInput);
+        // Zustand store에 저장
+        setSymptom(symptomList);
+        setAllergic(allergicList);
+
+        // 부모 컴포넌트로 이동
         onNext();
-        openModal();
     };
 
     return (
         <div>
             <Label
                 css="profileLabel"
-                text="김아기에 대해<br>
-더 알려줄 것이 있나요?"
+                text="김아기의 과거 증상과<br>
+진료 기록을 알려주세요"
             />
-            <div className={styles.etcArea}>
-                <textarea
-                    className={styles.etcInput}
-                    placeholder={`병원에서 미리 알아두면 좋을\n특이 사항이 있다면 적어주세요`}
-                    value={etcInput}
-                    onChange={(e) => setEtcInput(e.target.value)}
+            <div className={`${styles.listInputArea} mt-48 `}>
+                <Label
+                    text="과거에 진단받은 질환이 있나요?"
+                    css="profileDesc"
                 />
+                <div className="flex gap-12">
+                    <ListInput
+                        type="text"
+                        className="listInputForm"
+                        value={symptomInputValue}
+                        placeholder="ex:아토피"
+                        onChange={(e) => setSymptomInputValue(e.target.value)}
+                        onKeyDown={handleSymptomKeyDown}
+                    />
+                    <Circleline
+                        color="#729BED"
+                        className="symptomInputCheckbox"
+                        onClick={addSymptom}
+                    />
+                </div>
             </div>
-            <Button
-                css="nextBtn"
-                onClick={handleNext}
-                label="작성 완료"
-            ></Button>
+            <div className="mt-20 verticalFlexbox gap-8 mb-48">
+                {symptomList.map((symptom, index) => (
+                    <div key={index} className={styles.symptomList}>
+                        <Label css="profSymtomp" text={symptom} />
+                        <XIcon
+                            color="#222222"
+                            onClick={() => removeSymptom(index)}
+                        />
+                    </div>
+                ))}
+            </div>
+
+            <Label
+                text="약품이나 식품에<br>알러지가 있다면 적어주세요"
+                css="profileDesc"
+            />
+            <div className={`${styles.listInputArea}`}>
+                <div className="flex gap-12">
+                    <ListInput
+                        type="text"
+                        className="listInputForm"
+                        value={allergicInputValue}
+                        placeholder="ex:땅콩"
+                        onChange={(e) => setAllergicInputValue(e.target.value)}
+                        onKeyDown={handleAllergicKeyDown}
+                    />
+                    <Circleline
+                        color="#729BED"
+                        className="allergicInputCheckbox"
+                        onClick={addAllergic}
+                    />
+                </div>
+            </div>
+            <div className="mt-20 verticalFlexbox gap-8 mb-120">
+                {allergicList.map((allergic, index) => (
+                    <div key={index} className={styles.symptomList}>
+                        <Label css="profSymtomp" text={allergic} />
+
+                        <XIcon
+                            color="#222222"
+                            onClick={() => removeAllergic(index)}
+                        />
+                    </div>
+                ))}
+            </div>
+            <Button css="nextBtn" onClick={handleNext} label="다음"></Button>
         </div>
     );
 };
