+++ src/app/kid/register/chapter/Register.tsx
@@ -10,6 +10,7 @@ import Chapter4 from "./Chapter4";
 import Chapter5 from "./Chapter5";
 import axios from "axios";
 import useAuth from "@/hook/useAuth";
+import Chapter6 from "./Chapter6";
 
 const Register: React.FC = () => {
     const { getToken } = useAuth();
@@ -20,7 +21,7 @@ const Register: React.FC = () => {
         setToken(accessToken);
     }, []);
     const { chapter, nextChapter, goToChapter } = useChapter({
-        totalChapters: 5,
+        totalChapters: 6,
         onComplete: async () => {
             try {
                 const age = useProfileStore.getState().age;
@@ -104,13 +105,16 @@ const Register: React.FC = () => {
                 <Chapter2 onNext={nextChapter} goToChapter={() => {}} />
             )}
             {chapter === 3 && (
-                <Chapter3 onNext={nextChapter} goToChapter={goToChapter} />
+                <Chapter3 onNext={nextChapter} goToChapter={() => {}} />
             )}
             {chapter === 4 && (
-                <Chapter4 onNext={nextChapter} goToChapter={() => {}} />
+                <Chapter4 onNext={nextChapter} goToChapter={goToChapter} />
             )}
             {chapter === 5 && (
                 <Chapter5 onNext={nextChapter} goToChapter={() => {}} />
+            )}
+            {chapter === 6 && (
+                <Chapter6 onNext={nextChapter} goToChapter={() => {}} />
             )}
         </div>
     );
