+++ src/app/kid/register/kid.module.css
@@ -1,0 +1,102 @@
+.profileContainer {
+    display: flex;
+    flex-wrap: wrap;
+    gap: 8px;
+}
+
+.profileContainer_item {
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    height: 80px;
+    background-color: #ffffff;
+    color: var(--main-color);
+    border: 1px solid var(--main-color);
+    border-radius: 10px;
+    padding: 60px 0;
+    text-align: center;
+    font-size: var(--font-size-24);
+    font-weight: var(--font-weight-bold);
+    line-height: 1.2;
+}
+
+.inputArea {
+    display: flex;
+    flex-direction: column;
+    gap: 36px;
+    margin: 54px 0;
+}
+
+.listInputArea {
+    position: relative;
+}
+
+.symptomList {
+    width: 100%;
+    height: 40px;
+    position: relative;
+    display: flex;
+    align-items: center;
+    justify-content: space-between;
+    background-color: #f4f4f4;
+    border-radius: 8px;
+    padding: 0 20px;
+}
+
+.etcArea {
+    margin-top: 54px;
+    margin-bottom: 120px;
+}
+
+.etcInput {
+    width: 100%;
+    border: none;
+    border: 1px solid #d9d9d9;
+    border-radius: 10px;
+    min-height: 280px;
+    padding: 20px;
+}
+
+.etcInput::placeholder {
+    font-size: var(--font-size-16);
+    font-weight: var(--font-weight-medium);
+    color: #d9d9d9;
+}
+
+.modalHeader {
+    margin: 48px auto 40px auto;
+    text-align: center;
+    font-size: var(--font-size-24);
+    font-weight: var(--font-weight-semiBold);
+    line-height: 1.5;
+}
+
+.modalSubDesc {
+    text-align: center;
+    margin-bottom: 56px;
+    font-size: var(--font-size-24);
+    font-weight: var(--font-weight-medium);
+    color: #d9d9d9;
+}
+
+.genderSelection {
+    display: flex;
+    flex-direction: column;
+    gap: 12px;
+    justify-content: center;
+}
+
+.gender_item {
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    background-color: #ffffff;
+    color: var(--main-color);
+    border: 1px solid var(--main-color);
+    text-align: center;
+    font-size: var(--font-size-18);
+    font-weight: var(--font-weight-bold);
+    line-height: 1.2;
+    border-radius: 8px;
+    padding: 24px 0;
+}
